requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alpaca-trade-api==3.2.0
alpha-vantage==2.3.1
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
awscli==1.33.26
backtrader==1.9.78.123
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
deprecation==2.1.0
diskcache==5.6.3
docutils==0.16
et-xmlfile==1.1.0
fake-useragent==1.5.1
feedparser==6.0.11
finnhub-python==2.4.20
Flask==3.0.3
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.9.1
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
newsapi-python==0.2.7
nltk==3.8.1
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parse==1.20.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyee==11.1.0
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-html==0.10.0
rsa==4.7.2
s3transfer==0.10.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tqdm==4.66.4
twelvedata==1.2.15
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
w3lib==2.2.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.0.3
wh==1.2.0
yahoo-fin==0.8.9.1
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.2